7-consumer-data-storage/requirements.txt
pandas
ta
kafka-python
sqlalchemy
psycopg2-binary
python-dotenv
